update storage handling
